+++ unittests/CMakeLists.txt
@@ -7,13 +7,18 @@ set(${PROJECT_NAME}_SOURCES
     test_MatrixRoundTrip.cpp
     test_Sparse.cpp
     test_VectorRoundTrip.cpp
+    gtest-helpers.hpp
     )
 
 if (WIN32)
     add_definitions(/W4 /wd4251)
 else()
+    add_definitions(-Wall -Wextra)
 endif()
 
 add_executable(${PROJECT_NAME} ${${PROJECT_NAME}_SOURCES})
 
-target_link_libraries(${PROJECT_NAME} hdf5 hdf5_cpp GTest GTestMain)
+target_link_libraries(${PROJECT_NAME} GTest GTestMain)
+target_link_libraries(${PROJECT_NAME} ${HDF5_EXPORT_LIBRARIES})
+
+add_test(NAME ${PROJECT_NAME} COMMAND ${PROJECT_NAME} WORKING_DIRECTORY ${CMAKE_BINARY_DIR})
